Add new API function utf8_to_uv()
This is designed to replace the problematic utf8_to_uvchr(), which
is problematic.  Its behavior varies depending on if <utf8> warnings are
enabled or not, and no code in core actually takes that into account

If warnings are enabled:

    A zero return can mean both success or failure

        Hence a zero return must be disambiguated.  Success would come
        from the next character being a NUL.

    If failure, <retlen> will be -1, so can't be used to find where to
    start parsing again.

If disabled:

    Both the return and <retlen> will be usable values, but the return
    of the REPLACEMENT CHARACTER is ambiguous.  It could mean failure,
    or it could mean that that was the next character in the input and
    was successfully decoded.  It may very well not matter to you what
    the source of this particular value was.  It likely means a failure
    somewhere.  But there are occasions where you might care.

utf8_to_uv() solves these.  This commit includes a few changes to use
it, to show it works.  I have WIP that changes the rest of core to use
it.  I found that it makes coding simpler.

The new function returns true upon success; false on failure.  And it is
passed pointers to return the computed code point and byte length into.
These values always contain the correct information, regardless of if
the input is malformed or not.

It is easy to test for failure in a conditional and then to take
appropriate action.  However, most often it seems the appropriate action
is to use, going forward, the REPLACEMENT CHARACTER returned in failure
cases.

And if you don't care particularly if it succeeds or not, you just use
it without testing the result.  This happens when you are confident that
the input is well-formed, or say in converting a string for display.