ParseXS: re-indent two subs in Node.pm
The previous commit moved those subs from ParseXS.pm (where indentation
is 2 columns) to the new Node.pm file (where the indentation is 4
columns, because I created that file and so I get to choose). This
commit re-indents those subs to match the 4-indent.

Whitespace and line re-wrapping changes only.