ParseXS: refactor: XSUB signature-processing
The previous commit reduced the three separate XSUB signature-splitting
and initial processing branches into two. This commit reduces it to a
single common block of code. It also heavily refactors the initial
parameter processing code, and adds several new checks.

In outline: before this commit, the sig splitting and processing code
looked like:

    my @args;

    if (config_allow_argtypes) { # ANSI-style Sig
        if (can split using regex) {
            @args = /regex/g;
            for (@args) {
                # big loop to do lots of processing of all the extra
                syntactical features allowed with a full ANSI-style
                signature, i.e. IN/OUT, type, and length(foo),
                leaving the elements of @_ with all the extra stuff
                stripped.
            }
        }
        else {
           @args = split /,/, ...;
           Warn "Warning: cannot parse argument list";
        }
    }
    else { # K&R-style Sig
        @args = split /,/, ...;
        if (config_allow_inout) {
            # process IN/OUT etc
    }

    for (@args) {
        # do common K&R syntax processing of each parameter
    }

Following this commit, it now looks like:

    my @args;

    if (can split using regex) {
        @args = /regex/g;
    }
    else {
       @args = split /,/, ...;
       Warn "Warning: cannot parse argument list";
    }

    for (@args) {
        # big loop to do lots of processing of all the extra
        syntactical features allowed with a full ANSI-style
        signature, i.e. IN/OUT, type, and length(foo),
        leaving the elements of @_ with all the extra stuff
        stripped.

        $self->blurt("type not allowed")
            if defined $type && !$self{config_allow_argtypes};
        # ... and similar error checks ...
    }

    for (@args) {
        # do common K&R syntax processing of each parameter
    }

So under -noargtypes, instead of parsing a parameter by using a separate
block of code which doesn't know about the new syntactical features like
types etc, always parse using code which understands all the extra
stuff, but which then errors out with a specific error message if it
finds something disallowed under -noargtypes. This is better than
previously, where the forbidden thing was simply not understood by the
parsing code, and might trigger a confusing generic error, or be
silently accepted and cause malformed C code to be emitted.

The new errors are:

    "parameter type not allowed under -noargtypes")
    "length() pseudo-parameter not allowed under -noargtypes")
    "parameter IN/OUT modifier not allowed under -noinout")
    "Unparseable XSUB parameter: '$_'");
    "Default value not allowed on length() parameter '$name'"

That last one was already an error, but the text has been changed from:

      "Default value on length() argument: '$_'"

and it now does a $self->blurt() rather than a plain 'die', so the
correct line number is reported, and parsing can continue, looking for
further errors.

The next few commits will be moving all the code within that
"do common K&R syntax processing" block into the first 'for (@args)'
loop, eventually removing the second loop altogether.