ParseXS: refactor: centralise default arg parsing
There are two 'for (@args)' parsing loops, and the processing of the
'= default_expr' part of a parameter is currently split between the two
loops. This commit moves everything into the first loop: this is one of
a series of commits aimed at the complete elimination of the second loop.

It also stores the default value expression in its original form, and
now defers any quote-escaping to the code output stage.

In theory, no functional changes.