ParseXS: refactor/fix: move CLASS/THIS sig parsing
There are currently two 'for (@args)' parsing loops to parse an XSUB's
signature. Move the special-case code for handling a C++ XSUB's initial
THIS/CLASS parameter from before the second loop to before the first
loop.

This is one of a series of commits aimed at the complete elimination of
the second loop.

This commit also fixes a bug introduced a couple of commits ago that
caused the initial "THIS" or "CLASS" arg in this emitted code's error
message to be be skipped:

    if (items < N) croak("usage: foo(...)")

This commit also adds some tests for THIS/CLASS in usage message and
autocall parameters.