ParseXS: refactor: add Node::Sig class
Add a ExtUtils::ParseXS::Node::Sig class.

This is is a subclass of Node which is intended to hold all the
info parsed and extracted from an XSUB's signature plus any INPUT lines.
It mainly consists of an array of Node::Param objects, which hold info
about each parsed parameter, plus a hash which maps param names to
those param objects.

This commit does the basic work; subsequent commits will handle some
individual cases which are more complex.

The basic arrangement used to be like this:

The XSUB's signature (such as "a, OUT b, int c, d = 999") was split and
parsed, and the info about each parameter was stored in a collection of
hashes; for example:

    $self->{xsub_map_argname_to_in_out}{b} = 'OUT';
    $self->{xsub_map_argname_to_default}{d} = '999';
    etc.

In addition, a Node::Param object was created for each parameter which
has the *type* specified (such as as 'c' above), and pushed onto the
@ANSI_params array.

Then, each INPUT line was parsed, stored as a temporary Node::Param
object, then the as_code() method was called on that object, which
emitted the C declaration and initialisation code for that parameter.
as_code() would use the various {xsub_map_argname_to_foo} hashes
to lookup information needed for the code generation (such as the
default expression to emit).

Finally, as_code() was called on each param in @ANSI_params, to emit
declarations for params listed in the signature but not in an INPUT
line.

What this commit does is to replace all the various

    $self->{xsub_map_argname_to_foo}

fields with a single

    $self->{xsub_sig}

Node::Sig object; then it creates a Node::Param object for each
parameter in the signature and stores them in the Sig object.  This
means that the info about each param is now stored in a single
Node::Param object, rather than being stored across several hashes.

A couple of more tricksy xsub_map_argname_to_foo fields have been left
as-is for now, to be handled by individual commits to follow shortly.

The INPUT processing is left largely unchanged for now: it still, for
each INPUT line, creates a Node::Param object, calls as_code() on it,
and then immediately frees it. The difference being that as_code() now
looks up the info it needs from under $self->{xsub_sig} rather than from
lots of hashes.

The intention is that eventually those temporary INPUT objects won't be
discarded, but will instead be merged into the list of Node::Params in
the Node::Sig object. Code will then be emitted by iterating over the
params in the Sig object once all sig/INPUT parsing is complete. This
will be a significant step towards changing from an "emit code as we go
along" model into a "parse everything into an AST then walk the tree,
emitting code" model.