ParseXS: formally disallow dup THIS,CLASS
If an XSUB's name includes a class, it is treated as a C++ method
and an implicit initial 'THIS' or 'CLASS' parameter is added.

If the XSUB also includes an explicit THIS/CLASS variable, Bad Things
happen. There is in fact code in ParseXS.pm which detects 'THIS' in an
INPUT line and suppresses the emitting of a duplicate C var declaration;
but the generated C code is still wrong. For example, in

    X::Y::f(int THIS)

the C code will expect 2 passed args, for both the implicit and explicit
THIS, and the usage() message will be:

    croak_xs_usage(cv, "THIS, THIS")

As it happens, the changes from 3 commits ago to THIS/CLASS handling
turned such duplicate declarations into a compile-time error, .e.g.

    Error: duplicate definition of argument 'THIS' ignored in foo.xs, line N

This commit adds tests for those changes, and also removes the
now-redundant code from ParseXS which skips emitting a duplicate var
definition, as the dup will already have been reported as an error by
then.