ParseXS: THIS tests: add tests for no STDERR noise
For each set of tests in the THIS/CLASS block of tests, if there isn't
any explicit test for expected STDERR values, add a test that STDERR was
undef; i.e. explicitly test that there were no warnings or errors.

To have a suitable description for the new tests, extract out the common
prefix from the descriptions of each set of tests.

Also, add an extra boolean field to each test indicating that the
regex *shouldn't* match (this will be used in the next commit).