ParseXS: allow THIS/CLASS to be overridden
For a long time, ParseXS allowed the type of the implicit THIS variable
to be overridden. For example, while this:

    void
    X::Y::foo()

would cause this line of C to be emitted:

    X__Y * THIS = ....;

this :

    void
    X::Y::foo()
       int THIS

would cause this line of C to be emitted:

    int THIS = ....;

This wasn't documented, but might be useful and code out there might be
relying on it. This was broken by me a few commits ago, and this commit
restores the behaviour and adds tests.

There is a similar issue for new() methods, which create an implicit
CLASS parameter rather than THIS. In this case, ParseXS has *never*
handled it correctly. This:

    void
    X::Y::new()
       int CLASS

would cause *two* lines of C to be emitted:

    int    CLASS = ....;
    char * CLASS = ....;

This commit also fixes that, so now the type of both THIS *and* CLASS are
overrideable.

The fix works by adding a new field to the Node::Param object,
'soft_type', which is a type to be used unless explicitly overridden in
an INPUT line.

The code for this fix is a bit messy because the Node::Param objects in
the Node::Sig object, and the temporary Node::Param object created for
each line of an INPUT block, aren't yet unified.