ParseXS: add Node::Sig::usage_string() method.
Previously, to generate the string to output as C code like

    croak_xs_usage(cv,  "a, b,c = 0")

the code used a lexical var, @report_params, to build up a list of
parameter names, such as ('a', 'c = 0')  etc, as the parameters were
processed.

This commit gets rid of @report_params, and instead adds a new method,

    ExtUtils::ParseXS::Node::Sig::usage_string()

which examines the list of parameters stored in the Sig object
and deduces the usage string from that.

This is part of gradually moving towards a "build an AST *then* output
stuff based on that tree" model.