ParseXS: reimplement autocall arg calculation.
This commit changes how the argument string is generated which is output
as part of an autocall to a C function.

In particular:

- it deletes the ExtUtils::ParseXS::Utilities::C_func_signature()
  method;

- it deletes the test file t/110-assign_func_args.t which exercised that
  method;

- it adds a method called ExtUtils::ParseXS::Node::Sig::C_func_signature()
  which serves a similar purpose as the removed method, but in a
  different and more robust way;

- it adds several tests to t/001-basic.t which provide similar coverage
  to the deleted tests.

The former way of generating the arg string worked by: during XSUB
signature processing, a list of candidate arg names was accumulated, then
joined and stored as $self->{xsub_C_auto_function_signature} by
C_func_signature(), which handled OUT params and skipped any fake
THIS/CLASS arg. Then, If a C_ARGS keyword was then encountered, that
would overwrite the xsub_C_auto_function_signature string. Then when
INPUT lines were processed, any var declared as '&foo' would cause
something like this to be done:

    $self->{xsub_C_auto_function_signature} =~ s/\bfoo\b/&foo/.

This was not robust and could modify the C_ARGS overridden string.

The new method instead scans the parameter list in the Node::Sig  object
and builds the argument list from that *at the time it is needed*. Any
C_ARGS value is saved and used as an override.

In principle there are no functional changes, apart from bug fixes and
the deleting of a "not for public use" method.