ParseXS: reindent arg-processing code
Slightly widen the scope of a {...} block which encloses most of the
arg-splitting/processing code so that the block now encompasses *all*
the code, then re-indent the 'for (@args)' block within it, which was
was one indent too much due to earlier code refactoring.

Apart from the slight scope tweak, whitespace-only.